requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.26.1
anyio==4.3.0
async-timeout==4.0.3; python_version < "3.11"
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
brotli==1.1.0; platform_python_implementation == "CPython"
brotlicffi==1.1.0.0; platform_python_implementation != "CPython"
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0; platform_python_implementation != "CPython"
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6; platform_system == "Windows"
distro==1.9.0
ebooklib==0.18
exceptiongroup==1.2.1; python_version < "3.11"
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
groq==0.8.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
importlib-metadata==7.1.0
jinja2==3.1.3
jiter==0.4.0
langdetect==1.0.9
litellm==1.38.10
lxml==5.2.1
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
openai==1.30.3
packaging==24.0
promptdown==0.9.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22; platform_python_implementation != "CPython"
pydantic==2.7.1
pydantic-core==2.18.2
pydeeplx==1.0.7
pygments==2.17.2
pyparsing==3.1.2; python_version > "3.0"
python-dotenv==1.0.1
pyyaml==6.0.1
regex==2024.4.28
requests==2.32.2
rich==13.7.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.5
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
typing-extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
zipp==3.18.1
streamlit>=1.31
